Move Timer to fluid-core-utils
